+++ backend/src/main/java/org/vaccom/vcmgt/action/impl/NguoiDungActionImpl.java
@@ -166,6 +166,9 @@ public class NguoiDungActionImpl implements NguoiDungAction {
 		long coSoYTeId = bodyData.has(EntityConstant.COSOYTE_ID) ? bodyData.get(EntityConstant.COSOYTE_ID).longValue()
 				: 0;
 
+		long uyBanNhanDanId = bodyData.has(EntityConstant.UYBANNHANDAN_ID) ? bodyData.get(EntityConstant.UYBANNHANDAN_ID).longValue()
+				: 0;
+
 		if (Validator.isNull(hoVaTen)) {
 			throw new ActionException(MessageUtil.getVNMessageText("hovaten.empty"), HttpStatus.NOT_ACCEPTABLE.value());
 		}
@@ -173,6 +176,7 @@ public class NguoiDungActionImpl implements NguoiDungAction {
 		nguoiDung.setChucDanh(chucDanh);
 		nguoiDung.setCoSoYTeId(coSoYTeId);
 		nguoiDung.setDiaBanCoSoId(diaBanCoSoId);
+		nguoiDung.setUyBanNhanDanId(uyBanNhanDanId);
 		nguoiDung.setEmail(email);
 		nguoiDung.setHoVaTen(hoVaTen);
 //		 nguoiDung.setQuanTriHeThong(quanTriHeThong);