+++ backend/src/main/java/org/vaccom/vcmgt/constant/EntityConstant.java
@@ -81,7 +81,8 @@ public class EntityConstant {
 	public static final String BACSIKHAM = "BacSiKham";
 	public static final String SOMUIMOTCA = "SoMuiMotCa";
 	public static final String SOCATIEM = "SoCaTiem";
-	
+	public static final String UYBANNHANDAN_ID = "UyBanNhanDan_ID";
+
 	public static final String VAITRO = "VaiTro";
 
 }