+++ backend/src/main/java/org/vaccom/vcmgt/repository/UyBanNhanDanRepository.java
@@ -1,7 +1,10 @@
 package org.vaccom.vcmgt.repository;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import org.vaccom.vcmgt.entity.GiayDiDuong;
 import org.vaccom.vcmgt.entity.PhuongXa;
 import org.vaccom.vcmgt.entity.UyBanNhanDan;