+++ backend/src/main/java/org/vaccom/vcmgt/service/UyBanNhanDanService.java
@@ -1,8 +1,20 @@
 package org.vaccom.vcmgt.service;
 
-import org.springframework.stereotype.Service;
+import org.vaccom.vcmgt.entity.GiayDiDuong;
 import org.vaccom.vcmgt.entity.UyBanNhanDan;
-@Service
+
+import java.util.List;
+
 public interface UyBanNhanDanService {
-    public UyBanNhanDan findById (long id);
+    public UyBanNhanDan findById(long id);
+
+    public List<UyBanNhanDan> findAll(int page, int size);
+
+    public long countAll();
+
+    //Update
+    public UyBanNhanDan save(UyBanNhanDan uyBanNhanDan);
+
+    //Delete
+    public void delete(UyBanNhanDan uyBanNhanDan);
 }